Add files via upload
This project is a simple login and password manager that allows the user to create an account with a login and password, and then login to their account. Using basic cybersecurity techniques, I also programmed the password to be hidden from view when typed, and hashed making it difficult to be breached. When logging into the account, it hashes the inputted password and compares it to the already hashed password to determine whether the login is successful.